fix: move env files to respective directories
no real reason to put them in 'shared' if they aren't actually shared
